+++ eslint.config.mjs
@@ -8,4 +8,11 @@ const __dirname = path.dirname(__filename);
 
 export default defineConfig([{
     extends: [...nextCoreWebVitals],
+    rules: {
+        "react-hooks/set-state-in-effect": "off",
+        "react-hooks/refs": "off",
+        "react-hooks/preserve-manual-memoization": "off",
+        "react-hooks/static-components": "off",
+        "react-hooks/purity": "off",
+    },
 }]);